cloud_app_ver0/requirements-minimal.txt
Flask
Flask-CORS
google-cloud-speech
google-cloud-language
python-dotenv